refactor : separate api's hooks to queries and mutations
